fix: Defer processing of offline messages until after complete connection
Currently, a room is created for each received message, which might occur before the entire connection sequence is completed. This could lead to a situation where the RoomFeatures of the created room were incorrect for a Direct Message room if the server features were not already loaded. In that case catchup would fail since the room seemingly did not support MAM.